+++ code/classeine-lib/clsn/ui/amatista/AmatistaUISkin.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "clsn/ui/UISkin.h"
+
+namespace clsn::ui::amatista
+{
+    class AmatistaUISkin : public UISkin
+    {
+    public:
+        AmatistaUISkin();
+    };
+}